remove extra server layer in front of tower-lsp's server
I added it in as a CYA because tower-lsp is slightly unmaintained at the
moment, but the extra layer is added complexity we don't need right now.
There's a community fork of tower-lsp in the works right now, hopefully
we can just switch to that when it gains some traction.